backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]
